Move "generic" dialog visual-diff tests.
