+++ components/dialog/test/dialog-mixin.visual-diff.html
@@ -1,0 +1,35 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+	<link rel="stylesheet" href="../../../test/styles.css" type="text/css">
+	<script src="/node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
+	<script>
+		const preferNative = (window.location.search.indexOf('preferNative=false') === -1);
+		window.D2L = { DialogMixin: { preferNative: preferNative } };
+	</script>
+	<script type="module">
+		import '../../typography/typography.js';
+		import '../../button/button.js';
+		import '../dialog.js';
+	</script>
+	<title>d2l-dialog-mixin</title>
+	<meta name="viewport" content="width=device-width, initial-scale=1.0">
+	<meta http-equiv="X-UA-Compatible" content="ie=edge">
+	<meta charset="UTF-8">
+</head>
+<body class="d2l-typography">
+	<d2l-button id="open">Open Dialog</d2l-button>
+
+	<d2l-dialog id="dialog" title-text="Dialog Title" width="400">
+		<div id="top">Top</div>
+		<div>Line 1</div>
+		<div>Line 2</div>
+		<div>Line 3</div>
+		<div>Line 4</div>
+		<div id="bottom">Bottom</div>
+		<d2l-button slot="footer" primary>Primary</d2l-button>
+		<d2l-button slot="footer" id="cancel">Cancel</d2l-button>
+	</d2l-dialog>
+
+</body>
+</html>
